# docs/swagger.yaml
definitions:
  control-panel-service_internal_server_dto_response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  request.TestScenario:
    properties:
      execution_number_multi_agent:
        type: integer
      increase_agent_number:
        type: integer
      max_test_service_count:
        type: integer
      mother_service_id:
        type: integer
      name:
        type: string
      num_steps:
        type: integer
      test_category_id:
        type: integer
      test_service_config:
        $ref: '#/definitions/request.TestServiceConfigRequest'
    type: object
  request.TestScenarioPaginationRequest:
    properties:
      page:
        type: integer
      per_page:
        type: integer
    type: object
  request.TestScenarioUpdateRequest:
    properties:
      execution_number_multi_agent:
        type: integer
      id:
        type: integer
      increase_agent_number:
        type: integer
      max_test_service_count:
        type: integer
      mother_service_id:
        type: integer
      name:
        type: string
      num_steps:
        type: integer
      test_service_config:
        $ref: '#/definitions/request.TestServiceConfigRequest'
    type: object
...
  response.PaginatedTestScenario:
    properties:
      data:
        items:
          $ref: '#/definitions/response.TestScenario'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  response.TestScenarioResponseByID:
    properties:
      data:
        $ref: '#/definitions/response.TestScenario'
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/test-scenarios:
    post:
      consumes:
      - application/json
      description: Create a new test scenario with configuration
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.TestScenario'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Create a test scenario
      tags:
      - test-scenarios
  /api/v1/test-scenarios/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific test scenario by its ID
      parameters:
      - description: Test scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TestScenarioResponseByID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Get test scenario by ID
      tags:
      - test-scenarios
  /api/v1/test-scenarios/{id}/delete:
    post:
      consumes:
      - application/json
      description: Delete a specific test scenario by its ID.
      parameters:
      - description: Test scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Delete a test scenario.
      tags:
      - test-scenarios
  /api/v1/test-scenarios/{id}/pause:
    post:
      consumes:
      - application/json
      description: Retrieve a specific test scenario by its ID and pause the scenario.
      parameters:
      - description: Test scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Pause a test scenario.
      tags:
      - test-scenarios
  /api/v1/test-scenarios/{id}/resume:
    post:
      consumes:
      - application/json
      description: Retrieve a specific test scenario by its ID and resume the scenario.
      parameters:
      - description: Test scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Resume a test scenario.
      tags:
      - test-scenarios
  /api/v1/test-scenarios/{id}/start:
    post:
      consumes:
      - application/json
      description: Retrieve a specific test scenario by its ID and start the scenario.
      parameters:
      - description: Test scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Start a test scenario.
      tags:
      - test-scenarios
  /api/v1/test-scenarios/{id}/stop:
    post:
      consumes:
      - application/json
      description: Stop a specific test scenario by its ID.
      parameters:
      - description: Test scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Stop a test scenario.
      tags:
      - test-scenarios
  /api/v1/test-scenarios/search:
    post:
      consumes:
      - application/json
      description: Get test scenarios with pagination support
      parameters:
      - description: Pagination request with page and per_page
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.TestScenarioPaginationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedTestScenario'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Get paginated test scenarios
      tags:
      - test-scenarios
  /api/v1/test-scenarios/update:
    post:
      consumes:
      - application/json
      description: Update test scenario with configuration
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.TestScenarioUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Update a test scenario
      tags:
      - test-scenarios
swagger: "2.0"